+++ src/main/kotlin/com/example/backoffice/domain/admin/controller/AdminController.kt
@@ -3,10 +3,8 @@ package com.example.backoffice.domain.admin.controller
 import com.example.backoffice.domain.admin.dto.AdminSignUpRequest
 import com.example.backoffice.domain.admin.dto.AdminSignUpResponse
 import com.example.backoffice.domain.admin.service.AdminService
-import com.example.backoffice.domain.product.review.dto.ReviewResponse
 import com.example.backoffice.domain.user.dto.LoginRequest
 import com.example.backoffice.domain.user.dto.LoginResponse
-import com.example.backoffice.domain.user.service.UserService
 import com.example.backoffice.infra.security.CustomPreAuthorize
 import com.example.backoffice.infra.security.MemberPrincipal
 import com.example.backoffice.infra.security.MemberRole
@@ -30,23 +28,6 @@ class AdminController(
     @PostMapping("/login")
     fun login(@RequestBody loginRequest: LoginRequest): ResponseEntity<LoginResponse> {
         return ResponseEntity.status(HttpStatus.CREATED).body(adminService.login(loginRequest))
-    }
-
-    @GetMapping("/reviews")
-    fun getReviews(
-        @AuthenticationPrincipal principal: MemberPrincipal,
-        @RequestParam userId: Long?
-    ): ResponseEntity<List<ReviewResponse>> = preAuthorize.hasAnyRole(principal, setOf(MemberRole.ADMIN)) {
-        ResponseEntity.status(HttpStatus.OK).body(adminService.getReviews(userId))
-    }
-
-    @DeleteMapping("/reviews")
-    fun deleteReviews(
-        @AuthenticationPrincipal principal: MemberPrincipal,
-        @RequestBody reviewIds: List<Long>
-    ): ResponseEntity<Unit> = preAuthorize.hasAnyRole(principal, setOf(MemberRole.ADMIN)) {
-        adminService.deleteReviews(reviewIds)
-        ResponseEntity.status(HttpStatus.NO_CONTENT).build()
     }
 
     @DeleteMapping("/users/{userId}")
